Splits update and remove content into components on admin facing venue page
