+++ src/pages/Venue/AdminFacing/AdminView/index.styles.js
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const AdminView = styled.div`
+  .admin-controls-wrapper {
+    margin-bottom: 20px;
+
+    .admin-controls {
+      button:first-of-type {
+        margin-right: 10px;
+      }
+    }
+  }
+
+  @media screen and (min-width: ${({ theme }) => theme.screen.small}) {
+    .admin-controls-wrapper {
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+    }
+  }
+`;
