+++ src/pages/Venue/AdminFacing/DeleteView/index.styles.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const DeleteView = styled.div`
+  .delete-view-button {
+    display: block;
+    &:first-of-type {
+      margin-bottom: 10px;
+    }
+  }
+`;
