+++ src/pages/Venue/AdminFacing/index.styles.js
@@ -15,29 +15,4 @@ export const AdminSection = styled.section`
       background-color: ${({ theme }) => theme.colors.secondary};
     }
   }
-
-  .delete-view-button {
-    display: block;
-    &:first-of-type {
-      margin-bottom: 10px;
-    }
-  }
-
-  .admin-controls-wrapper {
-    margin-bottom: 20px;
-
-    .admin-controls {
-      button:first-of-type {
-        margin-right: 10px;
-      }
-    }
-  }
-
-  @media screen and (min-width: ${({ theme }) => theme.screen.small}) {
-    .admin-controls-wrapper {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-    }
-  }
 `;
